+++ src/main/java/com/hotel/database/jpa/JpaRequestRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * CRUD (Create, Read, Update, Delete) repository for requests
  */
-public interface JpaRequestRepository extends CrudRepository<AbstractRequest, Integer> {
+public interface JpaRequestRepository extends CrudRepository<AbstractRequest, Long> {
 }
