+++ src/main/java/com/hotel/service/item/intf/CartServiceInterface.java
@@ -14,14 +14,14 @@ public interface CartServiceInterface {
 
     void add(Cart cart);
 
-    Cart find(int id);
+    Cart find(long id);
 
-    void edit(@NotNull int id, boolean new_completed, int new_room, Set<Integer> new_item_ids);
+    void edit(@NotNull long id, boolean new_completed, int new_room, Set<Long> new_item_ids);
 
     void remove(Cart cart);
 
-    void remove(int id);
+    void remove(long id);
 
-    boolean exists(int id);
+    boolean exists(long id);
 
 }
