+++ src/main/java/com/hotel/service/request/MaintenanceRequestService.java
@@ -21,7 +21,7 @@ public class MaintenanceRequestService extends AbstractRequestService implements
     }
 
     @Override
-    public void edit(int ID, int new_room, long new_time, String new_notes, String new_commonRequestType) {
+    public void edit(long ID, int new_room, long new_time, String new_notes, String new_commonRequestType) {
         ((RequestWithNotesDatabase)database).edit(ID, new_room, new_time, new_notes, new_commonRequestType);
     }
 
@@ -46,12 +46,12 @@ public class MaintenanceRequestService extends AbstractRequestService implements
     }
 
     @Override
-    public void remove(int ID) {
+    public void remove(long ID) {
         database.remove(ID);
     }
 
     @Override
-    public AbstractRequest find(int ID) {
+    public AbstractRequest find(long ID) {
         return database.find(ID);
     }
 
@@ -61,7 +61,7 @@ public class MaintenanceRequestService extends AbstractRequestService implements
     }
 
     @Override
-    public boolean exists(int ID) {
+    public boolean exists(long ID) {
         return database.containsID(ID);
     }
