+++ src/main/java/com/hotel/service/request/intf/WakeUpRequestServiceInterface.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 
 public interface WakeUpRequestServiceInterface extends AbstractRequestServiceInterface {
 
-    void edit(int ID, int new_room, long new_time, long new_wakeup_time);
+    void edit(long ID, int new_room, long new_time, long new_wakeup_time);
 
     ArrayList<WakeUpRequest> findByWakeupTime(long wakeup_time);
